RaeburnOrchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raeburn-orchestrator"
version = "0.1.0"
description = "Raeburn Orchestrator pipeline demonstration"
authors = [{name = "Raeburn", email = "team@raeburn.com"}]
license = {text = "MIT"}
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
urls = {"Homepage" = "https://github.com/owner/RaeburnOrchestrator"}
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28",
    "aiofiles>=23.0",
    "aiosqlite>=0.20",
    "tenacity>=8.2",
    "sentence-transformers>=2.7",
    "pydantic>=2.6",
    "structlog>=24.1",
    "asyncpg>=0.29",
    "qdrant-client>=1.9",
    "opentelemetry-sdk>=1.35",
    "opentelemetry-exporter-otlp>=1.35",
    "fastapi>=0.111",
    "uvicorn>=0.27",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]
lint = [
    "ruff>=0.4",
    "mypy>=1.9",
    "types-aiofiles>=24.1"
]

[tool.setuptools.packages.find]
include = ["raeburn_brain", "agents", "core", "utils"]

[tool.setuptools]
package-dir = {"" = "."}

[project.scripts]
raeburn-orchestrator = "raeburn_brain.orchestrator.orchestrator:main"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
